reorganize to remove unnecessary dependencies, package into different barrel files
