+++ package.json
@@ -1,9 +1,9 @@
 {
   "name": "examma-ray",
-  "version": "1.0.136",
+  "version": "1.0.141",
   "description": "## Setup",
-  "main": "dist/examma-ray.js",
-  "types": "dist/examma-ray.d.ts",
+  "main": "dist/core.js",
+  "types": "dist/core.d.ts",
   "files": [
     "dist",
     "template",
