+++ src/core/ExamCurve.ts
