+++ src/core/assigned_exams.ts
@@ -3,7 +3,7 @@ import { parse_submission, SubmissionType } from './response/responses';
 import { ResponseKind } from './response/common';
 import { mk2html, mk2html_unwrapped } from './render';
 import { maxPrecisionString, renderPointsWorthBadge, renderScoreBadge, renderUngradedBadge } from "./ui_components";
-import { Exception, GraderMap } from './ExamGrader';
+import { Exception, GraderMap } from '../ExamGrader';
 import { QuestionGrader, GradingResult } from './QuestionGrader';
 import { isValidID, StudentInfo } from './specification';
 import { ExamComponentSkin } from './skins';
