+++ src/core/exam_components.ts
