+++ src/core/response/common.ts
