+++ src/core/response/fitb-drop.ts
