+++ src/core/response/multiple_choice.ts
