+++ src/core/response/util.ts
