+++ src/core/specification.ts
@@ -39,7 +39,7 @@ import { CHOOSE_ALL, Randomizer } from "./randomization";
 import { QuestionBank } from "./QuestionBank";
 import { ResponseKind } from "./response/common";
 import { ResponseSpecification } from "./response/responses";
-import { DEFAULT_SKIN, ExamComponentSkin, SkinChooser } from "./skins";
+import { ExamComponentSkin, SkinChooser } from "./skins";
 import { assert } from "./util";
 import { Exam, Question, Section } from "./exam_components";
