+++ src/core/util.ts
