--- src/examma-ray.ts
@@ -1,119 +1,0 @@
-export * from "./specification";
-
-export * from "./skins";
-
-export {
-  AssignedQuestion,
-  AssignedSection,
-  AssignedExam,
-} from "./assigned_exams";
-
-export {
-  Question,
-  Section,
-  Exam
-} from "./exam_components";
-
-export {
-  QuestionAnswer,
-  SectionAnswers,
-  ExamSubmission,
-  TrustedExamSubmission,
-  fillManifest
-} from "./submissions";
-
-export {
-  QuestionBank
-} from "./QuestionBank";
-
-export {
-  ExamUtils
-} from "./ExamUtils";
-
-export {
-  ResponseKind
-} from "./response/common";
-
-export {
-  ResponseSpecification,
-  SubmissionType,
-  ResponseHandler,
-  RESPONSE_HANDLERS
-} from "./response/responses";
-
-export {
-  CodeEditorSpecification,
-  CodeEditorSubmission
-} from "./response/code_editor";
-
-export {
-  FITBSpecification,
-  FITBSubmission
-} from "./response/fitb";
-
-export {
-  MCSpecification,
-  MCSubmission
-} from "./response/multiple_choice";
-
-export {
-  SASSpecification,
-  SASSubmission,
-  SASGroup,
-  SASItem
-} from "./response/select_a_statement";
-
-
-export {
-  ExamGenerator,
-  ExamGeneratorOptions
-} from "./ExamGenerator";
-
-export {
-  ExamGrader,
-  GraderMap,
-  Exception,
-  ExceptionMap
-} from "./ExamGrader";
-
-export {
-  QuestionGrader
-} from "./QuestionGrader";
-
-export {
-  FITBRegexGrader,
-  FITBRegexMatcher,
-  FITBRegexRubricItem,
-  FITBRegexGradingResult
-} from "./graders/FITBRegexGrader";
-
-export {
-  FreebieGrader,
-  FreebieGradingResult
-} from "./graders/FreebieGrader";
-
-export {
-  SimpleMCGrader,
-  SimpleMCGradingResult
-} from "./graders/SimpleMCGrader";
-
-export {
-  StandardSASGrader
-} from "./graders/StandardSASGrader";
-
-export {
-  SummationMCGrader,
-  SummationMCGradingResult
-} from "./graders/SummationMCGrader";
-
-export {
-  CodeWritingGrader,
-  CodeWritingRubricItem,
-  CodeWritingGradingResult
-} from "./graders/CodeWritingGrader";
-
-export {
-  GradingAssignmentSpecification,
-  QuestionGradingRecords,
-  QuestionGradingRecord
-} from "./grading/common";
