+++ src/graders/FreebieGrader.ts
@@ -1,6 +1,6 @@
-import { AssignedQuestion, GradedQuestion } from "../assigned_exams";
-import { ResponseKind, BLANK_SUBMISSION } from "../response/common";
-import { QuestionGrader, ImmutableGradingResult } from "../QuestionGrader";
+import { AssignedQuestion, GradedQuestion } from "../core/assigned_exams";
+import { ResponseKind, BLANK_SUBMISSION } from "../core/response/common";
+import { QuestionGrader, ImmutableGradingResult } from "../core/QuestionGrader";
 
 export type FreebieGradingResult = ImmutableGradingResult;
