+++ tsconfig.json
@@ -1,9 +1,13 @@
 {
   "include": [
-    "./src/examma-ray.ts",
+    "src/core.ts",
+    "src/ExamGenerator.ts",
+    "src/ExamGrader.ts",
+    "src/ExamUtils.ts",
+    "src/graders.ts",
     "./src/lib.d.ts",
     "./src/bin/init.ts",
-    "./src/grading/code-grader.ts",
+    "src/grading_interface/code-grader.ts",
     "src/frontend/frontend.ts",
     "src/test/lib.d.ts",
     "lib.d.ts"
@@ -20,7 +24,7 @@
     // "incremental": true,                   /* Enable incremental compilation */
     "target": "es2019",                          /* Specify ECMAScript target version: 'ES3' (default), 'ES5', 'ES2015', 'ES2016', 'ES2017', 'ES2018', 'ES2019', 'ES2020', or 'ESNEXT'. */
     "module": "commonjs",                     /* Specify module code generation: 'none', 'commonjs', 'amd', 'system', 'umd', 'es2015', 'es2020', or 'ESNext'. */
-    "lib": ["es2019", "DOM"],                             /* Specify library files to be included in the compilation. */
+    "lib": ["es2019"],                             /* Specify library files to be included in the compilation. */
     // "allowJs": true,                       /* Allow javascript files to be compiled. */
     // "checkJs": true,                       /* Report errors in .js files. */
     // "jsx": "preserve",                     /* Specify JSX code generation: 'preserve', 'react-native', or 'react'. */
